pyproject.toml
[tool.poetry]
name = "djangocourse"
version = "0.1.0"
description = ""
authors = ["Marcin Andrzejczak <amostori@op.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.0.1"
django-allauth = {extras = ["socialaccount"], version = "^65.3.1"}
dj-database-url = "^2.3.0"
psycopg = "^3.2.4"
django-anymail = {extras = ["mailgun"], version = "^12.0"}
django-widget-tweaks = "^1.5.0"
gunicorn = "^23.0.0"
whitenoise = {extras = ["brotli"], version = "^6.8.2"}


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"
django-browser-reload = "^1.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
